Separated commands into individual files
