+++ src/bot.js
@@ -1,7 +1,6 @@
 // Dependencies
 const config = require('../config.json');
 const Discord = require('discord.js');
-const utils = require('./utils.js');
 
 // Create new Discord client
 const bot = new Discord.Client({
@@ -12,122 +11,8 @@ const bot = new Discord.Client({
   ],
 });
 
-// ephemeral interaction response = only you see it
-
-// Commands
-const commands = {
-  // General
-  info: {
-    config: {
-      name: 'info',
-      description: 'Info about the bot',
-    },
-    handler: async (interaction) => {
-      const appInfo = await bot.application.fetch();
-      const embed = new Discord.MessageEmbed()
-        .setColor(config.embedColor)
-        .setTitle(`${appInfo.name} | Info`)
-        .setDescription('The same bot you know and love rewritten in Node.js to bring more control and slash commands')
-        .setThumbnail(appInfo.iconURL())
-        .setTimestamp()
-        .setFooter(`Created by ${appInfo.owner.tag}`, appInfo.owner.displayAvatarURL({ dynamic: true }));
-
-      await interaction.reply(embed);
-    },
-  },
-  ping: {
-    config: {
-      name: 'ping',
-      description: 'Pong!',
-    },
-    handler: async (interaction) => {
-      await interaction.reply('Command received, measuring latency...');
-      const msg = await interaction.channel.send('Ping...');
-      msg.edit(`**Pong!** \`${Date.now() - msg.createdAt}ms\``);
-    },
-  },
-  rng: {
-    config: {
-      name: 'rng',
-      description: 'Returns a random integer between min and max (both inclusive)',
-      options: [{
-        name: 'max',
-        description: 'Max value (Default: 6)',
-        type: 'INTEGER',
-      },
-      {
-        name: 'min',
-        description: 'Min value (Default: 1)',
-        type: 'INTEGER',
-      }],
-    },
-    handler: async (interaction) => {
-      const max = interaction.options[0]?.value || 6;
-      const min = interaction.options[1]?.value || 1;
-      if (min > max) {
-        interaction.reply(`${min} <= Result <= ${max}\nResult: Try again`);
-        return;
-      }
-      const result = Math.floor(Math.random() * (max - min + 1) + min);
-
-      interaction.reply(`${min} <= Result <= ${max}\nResult: ${result}`);
-    },
-  },
-  // Game management
-  songcheck: {
-    config: {
-      name: 'songcheck',
-      description: 'Send a message and ask participants to voice ownership of the given song',
-      options: [{
-        name: 'songid',
-        description: 'Song ID',
-        type: 'STRING',
-        required: true,
-      },
-      {
-        name: 'difficulty',
-        description: 'Difficulty',
-        type: 'STRING',
-        required: true,
-        choices: [{
-          name: 'beyond',
-          value: 'byn',
-        },
-        {
-          name: 'future',
-          value: 'ftr',
-        },
-        {
-          name: 'present',
-          value: 'prs',
-        },
-        {
-          name: 'past',
-          value: 'pst',
-        }],
-      }],
-    },
-    handler: async (interaction) => {
-      const appInfo = await bot.application.fetch();
-      const embed = new Discord.MessageEmbed()
-        .setColor(config.embedColor)
-        .setAuthor(appInfo.name, appInfo.iconURL({ dynamic: true }))
-        .setTitle('Song check')
-        .setDescription('Please check that you own this song')
-        .addFields(
-          { name: 'Title', value: interaction.options[0].value, inline: true },
-          { name: 'Pack', value: 'Arcaea', inline: true },
-          { name: 'Difficulty', value: interaction.options[1].value, inline: true },
-        )
-        .setTimestamp();
-
-      await interaction.reply('Command received, processing...', { ephemeral: true });
-      const msg = await interaction.channel.send(embed);
-      await msg.react('👍');
-      await msg.react('👎');
-    },
-  },
-};
+// Import commands
+const commands = require('./commands/index');
 const commandList = Object.keys(commands);
 
 bot.once('ready', async () => {
@@ -139,8 +24,8 @@ bot.once('ready', async () => {
   });
 
   // Create commands from commandList
-  commandList.forEach((commandName) => {
-    guild.commands.create(commands[commandName].config);
+  commandList.forEach(async (commandName) => {
+    await guild.commands.create(commands[commandName].config);
   });
 
   console.log(`Logged in as ${bot.user.tag}`);
@@ -150,7 +35,7 @@ bot.on('interaction', async (interaction) => {
   // Execute command handler
   if (!interaction.isCommand()) return;
   if (commandList.includes(interaction.commandName)) {
-    await commands[interaction.commandName].handler(interaction);
+    await commands[interaction.commandName].handler(bot, interaction);
     return;
   }
 });
