+++ src/commands/game/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+  songcheck: require('./songcheck'),
+};
