+++ src/commands/general/index.js
@@ -1,0 +1,5 @@
+module.exports = {
+  info: require('./info'),
+  ping: require('./ping'),
+  rng: require('./rng'),
+};
